Fix the generation of span ids to always be positive integers.
php_mt_rand() returns 32 bits of randomness even for 64 bit integers.
In the 64 bit case, a 0 is always right shifted into the 32nd bit.
In the 32 bit case, a right shift is dependent on the compiler
implementation. Generally, negative numbers will remain negative to make
bitshift arthmatic work.

+++ ext/opencensus_trace.c
@@ -250,7 +250,7 @@ static opencensus_trace_span_t *opencensus_trace_begin(zend_string *function_nam
         php_mt_srand(GENERATE_SEED());
     }
 #endif
-    span->span_id = (php_mt_rand() >> 1);
+    span->span_id = ((uint32_t) php_mt_rand()) >> 1;
 
     if (OPENCENSUS_TRACE_G(current_span)) {
         span->parent = OPENCENSUS_TRACE_G(current_span);
